+++ src/Common/strCoding.cpp
