+++ src/Common/strCoding.h
